requirements.txt
asgiref==3.8.1
boto3==1.34.84
botocore==1.34.84
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
jmespath==1.0.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
setuptools==68.2.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
wheel==0.41.2
